♻️ refactor: 누락된 수정 추가
